% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mdd.R
\name{compute_mdd}
\alias{compute_mdd}
\title{Compute Minimum Dietary Diversity (MDD)}
\usage{
compute_mdd(
  recall_data,
  id_col,
  food_item_col,
  life_stage,
  use_fct_db = FALSE,
  fg_map_data = NULL,
  fg_map_food = NULL,
  fg_map_col = NULL,
  breastfeeding_data = NULL,
  breastfeeding_col = NULL
)
}
\arguments{
\item{recall_data}{A data frame of recall data containing at least the ID
column and the food item column.}

\item{id_col}{Character. Name of the identifier column in \code{recall_data}
(e.g., \code{"survey_id"}).}

\item{food_item_col}{Character. Name of the column in \code{recall_data} containing food item names.}

\item{life_stage}{One of \code{"woman"}, \code{"man"}, \code{"adolescent"}, \code{"child"}.}

\item{use_fct_db}{Logical, default = FALSE. If TRUE, uses the package's
online FCT mapping database for food group classification.}

\item{fg_map_data}{Optional. A user-provided mapping data frame with
food items and food groups. Required if \code{use_fct_db = FALSE}.}

\item{fg_map_food}{Character. Column name in \code{fg_map_data} that matches food items.}

\item{fg_map_col}{Character. Column name in \code{fg_map_data} that contains the food group.}

\item{breastfeeding_data}{Optional. Data frame with ID and breastfeeding status (binary).
Only relevant if \code{life_stage = "child"}.}

\item{breastfeeding_col}{Character. Column name in \code{breastfeeding_data}
that contains breastfeeding status.}
}
\value{
A tibble with:
\itemize{
\item The ID column (as provided in \code{id_col})
\item One-hot encoded food group indicators
\item \verb{dds_<life_stage>} = Dietary Diversity Score
\item \verb{mdd_<life_stage>} = Binary indicator of meeting MDD
}
}
\description{
Computes Minimum Dietary Diversity (MDD) and Dietary Diversity Score (DDS)
for women, men, adolescents, or children, from cleaned dietary recall data.
}
\details{
\itemize{
\item For children: threshold is ≥5 if breastfeeding included, ≥4 otherwise.
\item For adults (woman, man, adolescent): threshold is always ≥5.
\item Warns if unmapped food items are found in recall data.
\item Final result for children returns the expected 7 food groups and adults returns the expected 10 food group, even if not mapped to any food item
\item While \code{Others} food group are part of the result, the column is not included in the summation of food groups for dds
}
}
\examples{
\dontrun{
# Using online FCT DB
result <- compute_mdd(
  recall_data = dietrecall_example$food_details,
  id_col = "survey_id",
  food_item_col = "desc_of_food",
  life_stage = "woman",
  use_fct_db = TRUE
)

# Using custom mapping
fg_map <- data.frame(
  item = c("Ugali", "Milk"),
  group = c("Cereals", "Dairy")
)

result <- compute_mdd(
  recall_data = dietrecall_example$food_details,
  id_col = "survey_id",
  food_item_col = "desc_of_food",
  life_stage = "child",
  use_fct_db = FALSE,
  fg_map_data = fg_map,
  fg_map_food = "item",
  fg_map_col = "group",
  breastfeeding_data = dietrecall_example$maintable,
  breastfeeding_col = "child_breastfed"
)
}

}
