% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ear_reference.R
\name{.ear_reference_table}
\alias{.ear_reference_table}
\title{--- Internal reference table (can be reused by other functions) ---}
\usage{
.ear_reference_table()
}
\description{
--- Internal reference table (can be reused by other functions) ---
}
\keyword{internal}
