% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iron_rni_bioavailability.R
\name{.get_iron_rni}
\alias{.get_iron_rni}
\title{Internal: Get iron RNI for a specific age, life group, and bioavailability}
\usage{
.get_iron_rni(age, life_group, bioavailability)
}
\description{
Internal: Get iron RNI for a specific age, life group, and bioavailability
}
\keyword{internal}
