% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rni_reference.R
\name{.get_rni}
\alias{.get_rni}
\title{Internal: Retrieve RNI for a specific nutrient, age, and life group}
\usage{
.get_rni(age, life_group, nutrient)
}
\description{
Internal: Retrieve RNI for a specific nutrient, age, and life group
}
\keyword{internal}
