% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinc_ear_bioavailability.R
\name{.get_zinc_ear}
\alias{.get_zinc_ear}
\title{Internal: Get zinc EAR for a specific age, life group, and bioavailability}
\usage{
.get_zinc_ear(age, life_group, bioavailability)
}
\description{
Internal: Get zinc EAR for a specific age, life group, and bioavailability
}
\keyword{internal}
