% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinc_rni_bioavailability.R
\name{.get_zinc_rni}
\alias{.get_zinc_rni}
\title{Internal: Get zinc RNI for a specific age, life group, and bioavailability}
\usage{
.get_zinc_rni(age, life_group, bioavailability)
}
\description{
Internal: Get zinc RNI for a specific age, life group, and bioavailability
}
\keyword{internal}
