% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iron_rni_bioavailability.R
\name{.iron_rni_bioavailability}
\alias{.iron_rni_bioavailability}
\title{Internal: Iron RNI values adjusted for bioavailability}
\usage{
.iron_rni_bioavailability()
}
\description{
Returns RNI values (mg/day) for iron across life groups, adjusted by
bioavailability (low = 5\%, moderate = 10\%, high = 15\%).
}
\keyword{internal}
