% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinc_ear_bioavailability.R
\name{.zinc_ear_bioavailability}
\alias{.zinc_ear_bioavailability}
\title{Internal: Zinc EAR values adjusted for bioavailability}
\usage{
.zinc_ear_bioavailability()
}
\description{
Returns a nested list of zinc EARs (mg/day) based on life group and bioavailability.
Bioavailability levels: "low" (5\%), "moderate" (10\%), "high" (15\%).
Age is assumed to be in years for adults and months for children.
}
\keyword{internal}
