% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinc_rni_bioavailability.R
\name{.zinc_rni_bioavailability}
\alias{.zinc_rni_bioavailability}
\title{Internal: Zinc RNI values adjusted for bioavailability}
\usage{
.zinc_rni_bioavailability()
}
\description{
Returns RNI values (mg/day) for zinc across life groups, adjusted by
bioavailability (low = 5\%, moderate = 10\%, high = 15\%).
}
\keyword{internal}
