% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello_world}
\alias{hello_world}
\title{Hello World Function}
\usage{
hello_world()
}
\value{
A character string.
}
\description{
This is a placeholder function to ensure the package passes R CMD check.
Delete once you add real functions.
}
